Update FindDoctorPage.ts
I added a dynamic navigate method that allows us to navigate to additional results pages. String interpolation is used to pass the search term, location, and page number into the url. Please note that the url must be all clumped in one long line of code to function properly. I tried making it look better, but had errors within the test because of it.